requirements.txt
phidata
python-dotenv
yfinance
packaging
ddgs
fastapi
uvicorn
groq
python-multipart